src/server/requirements.txt
numpy==1.18
biopython==1.76
mmh3==2.5.1
bitarray==1.2.0
scikit-learn==0.20
scikit-allel==1.2.1
pyfaidx==0.5.7
pysam==0.15.3
PyVCF==0.6.8